pyproject.toml
[project]
name = "lite-star"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "advanced-alchemy (>=1.1.1,<2.0.0)",
    "msgspec (>=0.19.0,<0.20.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "litestar[standart] (>=2.15.2,<3.0.0)",
    "litestar-granian (>=0.12.2,<0.13.0)",
    "litestar-asyncpg (>=0.3.0,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
